Tidy depth CI tests: merge into main suites, consolidate depth folder
Depth tests with a natural home in the existing suites move there; only
depth-domain tests (calibration, metrics protocol, val plotting, dataset
GT semantics) stay under tests/depth/.

- test_export.py -> test_exports.py: the two ONNX tests collapse into one
  dynamic export (nms forced off + output names + dynamic dims in a single
  export), renamed test_export_onnx_*/test_export_imx_* so the --export-env
  partition assigns them correctly
- test_loss.py, test_head.py, test_results.py, test_cfg.py, test_no_env.py,
  test_dataloader_shutdown.py -> test_python.py next to their existing
  counterparts (semantic loss, nn modules, results, cfg, dataloader tests)
- test_engine.py -> test_engine.py (hermetic synthetic-dataset cycle)
- streamlit task-order check -> test_solutions.py
- test_validator.py merged into test_metrics.py; test_pipeline.py +
  test_transforms.py consolidated as test_dataset.py

Deleted: tests/depth/{test_cfg,test_no_env,test_dataloader_shutdown,
test_engine,test_export,test_head,test_loss,test_results,test_pipeline,
test_transforms,test_validator}.py (contents redistributed);
test_depth_task_trainers_importable (the engine cycle already exercises
task_map dispatch to the depth Trainer/Validator/Predictor); the separate
coreml head-upsample and static-ONNX-export tests (folded into their
onnx counterparts). Net -88 lines.
